feat: 기존의 로그인 화면과 통합(with ChatGPT)
