+++ src/main/resources/templates/yjshop/kakaoLogin.html
@@ -31,10 +31,8 @@
 </head>
 <body>
 <div class="container">
-  <img src="/yjshop.png" alt="YJ Shop" class="logo">
-  <a th:href="${login_url}">
-    <img src="/kakao_login.png" alt="Kakao Login" class="login-button">
-  </a>
+  <img src="/yjshop.png" alt="YJ Shop"> class="logo">
+
 </div>
 </body>
 </html>
