+++ src/main/resources/templates/yjshop/user/login.html
@@ -26,16 +26,22 @@
       flex-direction: column;
     }
 
-    #loginInterceptor-page .form-title {
-      text-align: center;
-      font-size: 20px;
+    /* 로고 이미지 영역 */
+    #loginInterceptor-page .login-logo {
+      display: flex;
+      justify-content: center;
       margin-bottom: 25px;
+    }
+
+    #loginInterceptor-page .login-logo img {
+      max-width: 150px; /* 약 25% 축소 */
+      height: auto;
     }
 
     #loginInterceptor-page .input-group {
       display: flex;
       flex-direction: column;
-      margin-bottom: 30px; /* 입력칸 사이 넉넉한 간격 */
+      margin-bottom: 30px;
     }
 
     #loginInterceptor-page .input-group label {
@@ -74,13 +80,33 @@
     #loginInterceptor-page button {
       background-color: #ddd;
     }
+
+    /* 카카오 로그인 버튼 */
+    #loginInterceptor-page .kakao-login {
+      display: block;
+      margin-top: 10px;
+      width: 100%;
+    }
+
+    #loginInterceptor-page .kakao-login img {
+      width: 100%;
+      height: 45px; /* 기존보다 살짝 낮게 */
+      object-fit: cover;
+      display: block;
+    }
+
   </style>
 </head>
 <body>
 <div id="loginInterceptor-page">
   <div class="loginInterceptor-form-container">
+
+    <!-- 로고 이미지 -->
+    <div class="login-logo">
+      <img src="/yjshop.png" alt="YJ Shop">
+    </div>
+
     <form th:action="@{/view/login}" th:object="${memberRequestDto}" method="post">
-      <div class="form-title">로그인</div>
 
       <!-- 이메일 입력 -->
       <div class="input-group">
@@ -95,8 +121,8 @@
       <div class="input-group">
         <label for="password">비밀번호</label>
         <input type="password" id="password" th:field="*{password}" placeholder="비밀번호를 입력하세요">
-        <div th:if="${#fields.hasErrors('password')}" th:errors="*{password}"
-             class="warning-message">Password Error
+        <div th:if="${#fields.hasErrors('password')}" th:errors="*{password}" class="warning-message">
+          Password Error
         </div>
       </div>
 
@@ -105,6 +131,12 @@
 
     <!-- 회원가입 버튼 -->
     <button th:onclick="|location.href='@{/view/registerform}'|">회원 가입</button>
+
+    <!-- 카카오 로그인 버튼 -->
+    <a th:href="${login_url}" class="kakao-login">
+      <img src="/kakao_login.png" alt="Kakao Login">
+    </a>
+
   </div>
 </div>
 </body>
